linux: Debugging and troubleshooting the ip header protocol

+++ include/NetworkSnifferManager.h
@@ -1,14 +1,15 @@
+#include <cstdint>
 #include <unistd.h>
 #include <netinet/ip.h>
 #include <netinet/ip_icmp.h>
 #include <netinet/udp.h>
 #include <netinet/tcp.h>
-#include <netinet/in.h>
+
 
 class NetworkSnifferManager {
 
 public:
-	NetworkSnifferManager(int BufferSize);
+	NetworkSnifferManager(size_t BufferSize);
 	~NetworkSnifferManager();
 	void OpenConnection();
 	void ReadPacket(unsigned char *read_buffer, int dataSize);
@@ -19,7 +20,7 @@ public:
 	sockaddr_in6 sockAddressIPv6Source, sockAddressIPv6Dest;
 	unsigned char *readBuffer;
 
-	int bufferSize;
+	size_t bufferSize;
 	int socketFD;
 	int tcp;
 	int udp;
